* MOD: updated PKGBUILD for easier installation * MOD: updated compile.sh to not include files not present anymore

+++ compile.sh
@@ -1,10 +1,8 @@
 #!/bin/sh
-rm -rf src/
+[[ -d "src" ]] && rm -rf src/
 [[ -f "config.def.h" ]] && rm config.def.h
 [[ "$(hostname)" == "uh-x200s" ]] && cp config.def.h.x200s config.def.h
 [[ "$(hostname)" == "uh-x230" ]] && cp config.def.h.x230 config.def.h
-[[ "$(hostname)" == "uh-lin" ]] && cp config.def.h.work config.def.h
 updpkgsums
 makepkg -fi --noconfirm
-rm -rf dwm/
-rm config.def.h
+[[ -d "dwm" ]] && rm -rf dwm/
